Fix foreign-key type mismatch breaking saves/deletes in production
Saving a file's code, deleting a share, restoring a version, and
several other actions all compare a related model's foreign key
against a parent's id using ===, e.g.
$file->project_id === $project->id. Locally, PDO/mysqlnd returns
native PHP ints for integer columns, so this always worked. On the
production host, the PHP/MySQL driver build returns those same
columns as strings — "3" === 3 is false in PHP — so every one of
these checks silently 404'd/403'd. This is why saving code in the
editor appeared to "go blank" on refresh: the save request never
persisted, so the refresh just reloaded the original unsaved content.

Fixed at the model layer rather than patching each comparison site:
every foreign-key column across Project, ProjectFile, Share,
ProjectVersion, ProjectVersionFile, Comment, Media, Folder, BrandKit,
ViewLog, AgreementSignature, and User is now explicitly cast to
'integer', so Eloquent guarantees a real PHP int regardless of what
the DB driver returns. One spot (CommentController's cross-org check)
uses a raw query builder ->value() call that bypasses model casts
entirely, so it's cast explicitly inline instead.

+++ app/Http/Controllers/CommentController.php
@@ -135,9 +135,12 @@ class CommentController extends Controller
         // Share has an organization-scoping global scope, so $comment->share silently
         // resolves to null (not 403) when the authenticated user belongs to a different
         // org than the comment's share — query around the scope to get a real comparison.
-        $shareOrgId = Share::withoutGlobalScope('organization')->whereKey($comment->share_id)->value('organization_id');
+        // ->value() is a raw query builder call, not a hydrated model, so it bypasses
+        // Share's own attribute casts — cast explicitly rather than relying on whatever
+        // type the DB driver happens to return for an integer column.
+        $shareOrgId = (int) Share::withoutGlobalScope('organization')->whereKey($comment->share_id)->value('organization_id');
 
-        abort_unless($user->organization_id === $shareOrgId, 403);
+        abort_unless((int) $user->organization_id === $shareOrgId, 403);
     }
 
     private function authorizeDeleteAccess(Request $request, Comment $comment): void
